% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExchangeRates.R
\name{getAllExchangeRates}
\alias{getAllExchangeRates}
\title{Get all exchange rates for a symbol}
\usage{
getAllExchangeRates(assetIdBase, invert = FALSE, filter_asset_id = NULL)
}
\arguments{
\item{filter_asset_id}{}
}
\value{

}
\description{
Get all exchange rates for a symbol
}
\examples{
getAllExchangeRates("BTC")
}
