% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Markets.R
\name{getAssetIcons}
\alias{getAssetIcons}
\title{Get icon URLs for all listed Assets}
\usage{
getAssetIcons(sizeText)
}
\arguments{
\item{sizeText}{}
}
\value{

}
\description{
Get icon URLs for all listed Assets
}
\examples{
getAssetIcons(32)
}
