% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Markets.R
\name{getAssets}
\alias{getAssets}
\title{Get all Assets listed by CoinAPI}
\usage{
getAssets(assetId = NULL)
}
\arguments{
\item{assetId}{}
}
\value{

}
\description{
Get all Assets listed by CoinAPI
}
\examples{
getAssets()
getAssets("BTC")
}
