% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Markets.R
\name{getCryptoSymbols}
\alias{getCryptoSymbols}
\title{Get all symbols listed by CoinAPI - named to not clash with Quantmod}
\usage{
getCryptoSymbols(
  filterSymbolId = NULL,
  exchangeId = NULL,
  filterAssetId = NULL
)
}
\arguments{
\item{assetId}{}
}
\value{

}
\description{
Get all symbols listed by CoinAPI - named to not clash with Quantmod
}
\examples{
getCryptoSymbols()
getCryptoSymbols(filterSymbolId = "BTC")
getCryptoSymbols(filterSymbolId = "BTC", filterAssetId = "USD")
getCryptoSymbols(filterSymbolId = "BTC", filterAssetId = "USD", exchangeId = "FTX")
}
