% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Markets.R
\name{getExchangeIcons}
\alias{getExchangeIcons}
\title{Get URLs of all exchange icons by size}
\usage{
getExchangeIcons(sizeText)
}
\arguments{
\item{sizeText}{}
}
\value{

}
\description{
Get URLs of all exchange icons by size
}
\examples{
getExchangeIcons(32)
}
