% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExchangeRates.R
\name{getExchangeRate}
\alias{getExchangeRate}
\title{Get exchange rate for a base/quote pair}
\usage{
getExchangeRate(assetIdBase, assetIdQuote, time = NULL)
}
\arguments{
\item{time}{}
}
\value{

}
\description{
Get exchange rate for a base/quote pair
}
\examples{
getExchangeRate("BTC", "USD")
}
