% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Markets.R
\name{getExchanges}
\alias{getExchanges}
\title{Get list of all exchanges provided by CoinAPI}
\usage{
getExchanges(exchangeId = NULL)
}
\arguments{
\item{exchangeId}{}
}
\value{

}
\description{
Get list of all exchanges provided by CoinAPI
}
\examples{
getExchanges()
getExchanges("FTX")
}
