% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExchangeRates.R
\name{getHistoricalExchangeRatePeriods}
\alias{getHistoricalExchangeRatePeriods}
\title{Get valid periods for historical exchange rates queries}
\usage{
getHistoricalExchangeRatePeriods()
}
\value{

}
\description{
Get valid periods for historical exchange rates queries
}
\examples{
getHIstoricalExchangeRatePeriods()
}
