% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExchangeRates.R
\name{getHistoricalExchangeRates}
\alias{getHistoricalExchangeRates}
\title{Get the historical exchange rates between two assets as an XTS object}
\usage{
getHistoricalExchangeRates(
  assetIdBase,
  assetIdQuote,
  periodId,
  timeStart,
  timeEnd,
  limit = 100
)
}
\arguments{
\item{limit}{}
}
\value{

}
\description{
Get the historical exchange rates between two assets as an XTS object
}
\examples{
getHistoricalExchangeRates("BTC", "USD", "1HRS", "2016-01-01T00:00:00", "2016-02-01T00:00:00", limit = 25)
}
