% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OHLCV.R
\name{getHistoricalOHLCV}
\alias{getHistoricalOHLCV}
\title{Get OHLCV timeseries data returned in time ascending order. Data can be requested by the period and symbol or asset pairs.
If asset pairs are used then timeseries in the response will contain data consolidated across all most legitimate SPOT
markets which contain both assets in any order..}
\usage{
getHistoricalOHLCV(
  assetIdBase,
  assetIdQuote = NULL,
  periodId,
  timeStart,
  timeEnd = NULL,
  includeEmptyItems = FALSE,
  limit = 100
)
}
\arguments{
\item{limit}{}
}
\value{

}
\description{
Get OHLCV timeseries data returned in time ascending order. Data can be requested by the period and symbol or asset pairs.
If asset pairs are used then timeseries in the response will contain data consolidated across all most legitimate SPOT
markets which contain both assets in any order..
}
\examples{
getHistoricalOHLCV("BTC", "USD", periodId = "1MIN", timeStart = "2016-01-01T00:00:00", limit = 25)
}
