% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Orderbook.R
\name{getHistoricalOrderbook}
\alias{getHistoricalOrderbook}
\title{Get historical order book snapshots for a specific symbol within time range, returned in time ascending order.}
\usage{
getHistoricalOrderbook(
  symbol,
  limitLevels = 0,
  timeStart,
  timeEnd = NULL,
  limit = 100
)
}
\arguments{
\item{limit}{}
}
\value{

}
\description{
Get historical order book snapshots for a specific symbol within time range, returned in time ascending order.
}
\examples{
getHistoricalOrderbook(symbol = "BITSTAMP_SPOT_BTC_USD", timeStart = "2016-01-01T00:00:00", timeEnd = "2016-01-01T00:10:00", limitLevels = 1, limit = 25)
}
