% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quotes.R
\name{getHistoricalQuotes}
\alias{getHistoricalQuotes}
\title{Get historical quote updates within requested time range, returned in time ascending order.}
\usage{
getHistoricalQuotes(symbol, timeStart, timeEnd = NULL, limit = 100)
}
\arguments{
\item{limit}{}
}
\value{

}
\description{
Get historical quote updates within requested time range, returned in time ascending order.
}
\examples{
getHistoricalQuotes(symbol = "BITSTAMP_SPOT_BTC_USD", timeStart = "2016-01-01T00:00:00", timeEnd = "2016-01-01T00:10:00", limit = 25)
}
