% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trades.R
\name{getHistoricalTrades}
\alias{getHistoricalTrades}
\title{Get historical transactions from specific symbol, returned in time ascending order.}
\usage{
getHistoricalTrades(
  symbol,
  timeStart,
  timeEnd = NULL,
  limit = 100,
  include = FALSE
)
}
\arguments{
\item{include}{}
}
\value{

}
\description{
Get historical transactions from specific symbol, returned in time ascending order.
}
\examples{
getHistoricalTrades(symbol = "BITSTAMP_SPOT_BTC_USD", timeStart = "2016-01-01T00:00:00", limit = 25)
getHistoricalTrades(symbol = "BITSTAMP_SPOT_BTC_USD", timeStart = "2016-01-01T00:00:00", timeEnd = "2016-01-01T00:10:00", limit = 25)
}
