% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OHLCV.R
\name{getLatestOHLCV}
\alias{getLatestOHLCV}
\title{Get OHLCV latest timeseries data returned in time descending order. Data can be requested by the period and symbol or asset pairs.
If asset pairs are used then timeseries in the response will contain data consolidated across all most legitimate SPOT markets which
contain both assets in any order.}
\usage{
getLatestOHLCV(
  assetIdBase,
  assetIdQuote = NULL,
  periodId,
  includeEmptyItems = FALSE,
  limit = 100
)
}
\arguments{
\item{limit}{}
}
\value{

}
\description{
Get OHLCV latest timeseries data returned in time descending order. Data can be requested by the period and symbol or asset pairs.
If asset pairs are used then timeseries in the response will contain data consolidated across all most legitimate SPOT markets which
contain both assets in any order.
}
\examples{
getLatestOHLCV("BITSTAMP_SPOT_BTC_USD", "1MIN", limit = 25)
getLatestOHLCV("BTC", "USD", periodId = "1MIN", limit = 25)
}
