% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Orderbook.R
\name{getOrderbook}
\alias{getOrderbook}
\title{Get current order book snapshot for all or a specific symbol.}
\usage{
getOrderbook(symbol = NULL, filterSymbol = NULL, limitLevels = 0)
}
\arguments{
\item{limitLevels}{}
}
\value{

}
\description{
Get current order book snapshot for all or a specific symbol.
}
\examples{
getOrderbook(symbol = "BITSTAMP_SPOT_BTC_USD", limitLevels = 1)
getOrderbook(filterSymbol = "BTC")
}
