% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrderbookL3.R
\name{getOrderbookL3}
\alias{getOrderbookL3}
\title{Get current order book snapshot for all or a specific symbol.}
\usage{
getOrderbookL3(symbol = NULL, filterSymbol = NULL, limitLevels = 0)
}
\arguments{
\item{limitLevels}{}
}
\value{

}
\description{
Get current order book snapshot for all or a specific symbol.
}
\examples{
getOrderbookL3(filterSymbol = "BTC", limitLevels = 1)
getOrderbookL3(symbol = "COINBASE_SPOT_BCH_USD", limitLevels = 1)
}
