% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quotes.R
\name{getQuotes}
\alias{getQuotes}
\title{Get current quotes for all symbols or for a specific symbol.}
\usage{
getQuotes(symbol = NULL, filterSymbol = NULL)
}
\arguments{
\item{filterSymbol}{}
}
\value{

}
\description{
Get current quotes for all symbols or for a specific symbol.
}
\examples{
getQuotes()
getQuotes(symbol = "BITSTAMP_SPOT_BTC_USD")
getQuotes(filterSymbol = "BTC")
}
