% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trades.R
\name{getTrades}
\alias{getTrades}
\title{Get latest trades from all symbols up to 1 minute ago or get latest trades from a specific symbol without time limitation. Latest data is always returned in time descending order.}
\usage{
getTrades(symbol = NULL, limit = 100, filterSymbol = NULL, include = FALSE)
}
\arguments{
\item{include}{}
}
\value{

}
\description{
Get latest trades from all symbols up to 1 minute ago or get latest trades from a specific symbol without time limitation. Latest data is always returned in time descending order.
}
\examples{
getTrades()
getTrades(filterSymbol = "BTC", limit = 25)
getTrades(symbol = "BITSTAMP_SPOT_BTC_USD", limit = 25)
}
