% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{listToStringArray}
\alias{listToStringArray}
\title{Convert Lists to url-parameter compliant string arrays}
\usage{
listToStringArray(l)
}
\arguments{
\item{l}{}
}
\value{

}
\description{
Convert R lists to url-parameter compliant string arrays
}
\examples{
listToStringArray(list("Foo", "Bar"))
}
